requirements.txt
asgiref==3.9.2
attrs==25.3.0
autopep8==2.3.2
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
clize==5.0.2
Django==4.0.8
django-tastypie==0.15.1
docutils==0.22.2
Flask==3.1.2
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
od==2.0.2
pycodestyle==2.14.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-mimeparse==2.0.0
PyYAML==6.0.3
requests==2.32.5
sigtools==4.0.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
Werkzeug==3.1.3
